# projects/ranking-api-productization-showcase/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "model_loaded": {
            "title": "Model Loaded",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "model_loaded"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ModelSchemaResponse": {
        "properties": {
          "feature_names": {
            "items": {
              "type": "string"
            },
            "title": "Feature Names",
            "type": "array"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "required": [
          "feature_names"
        ],
        "title": "ModelSchemaResponse",
        "type": "object"
      },
      "PlayerRank": {
        "properties": {
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "player_id",
          "score",
          "rank"
        ],
        "title": "PlayerRank",
        "type": "object"
      },
      "PlayerRecord": {
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Features",
            "type": "object"
          },
          "player_id": {
            "minLength": 1,
            "title": "Player Id",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "features"
        ],
        "title": "PlayerRecord",
        "type": "object"
      },
      "PlayerScore": {
        "properties": {
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "player_id",
          "score"
        ],
        "title": "PlayerScore",
        "type": "object"
      },
      "RankResponse": {
        "properties": {
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/PlayerRank"
            },
            "title": "Rankings",
            "type": "array"
          }
        },
        "required": [
          "rankings"
        ],
        "title": "RankResponse",
        "type": "object"
      },
      "ScoreRequest": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/PlayerRecord"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
      "ScoreResponse": {
        "properties": {
          "scores": {
            "items": {
              "$ref": "#/components/schemas/PlayerScore"
            },
            "title": "Scores",
            "type": "array"
          }
        },
        "required": [
          "scores"
        ],
        "title": "ScoreResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ranking-api-productization-showcase",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/model/schema": {
      "get": {
        "operationId": "model_schema_model_schema_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSchemaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "Model not loaded"
          }
        },
        "summary": "Model Schema"
      }
    },
    "/predict": {
      "post": {
        "operationId": "predict_endpoint_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid model features"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Model not loaded"
          }
        },
        "summary": "Predict Endpoint"
      }
    },
    "/rank": {
      "post": {
        "operationId": "rank_endpoint_rank_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid model features"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Model not loaded"
          }
        },
        "summary": "Rank Endpoint"
      }
    },
    "/score": {
      "post": {
        "operationId": "score_endpoint_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid model features"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Model not loaded"
          }
        },
        "summary": "Score Endpoint"
      }
    }
  }
}